+++ domain/src/commonMain/kotlin/io/github/alelk/tgvd/domain/video/VideoInfo.kt
@@ -19,6 +19,22 @@ data class VideoInfo(
     val thumbnails: List<Thumbnail> = emptyList(),
     val description: String? = null,
     val viewCount: Long? = null,
+    val availableFormats: List<Format> = emptyList(),
+    val actualFormat: Format? = null,
 ) {
     data class Thumbnail(val url: Url, val width: Int?, val height: Int?)
+
+    data class Format(
+        val formatId: String,
+        val extension: String,
+        val width: Int? = null,
+        val height: Int? = null,
+        val fps: Double? = null,
+        val tbr: Double? = null,
+        val vcodec: String? = null,
+        val acodec: String? = null,
+        val formatNote: String? = null,
+        val filesize: Long? = null,
+        val filesizeApprox: Long? = null,
+    )
 }