+++ domain/src/commonMain/kotlin/io/github/alelk/tgvd/domain/video/VideoInfoCache.kt
@@ -3,5 +3,6 @@ package io.github.alelk.tgvd.domain.video
 interface VideoInfoCache {
     suspend fun get(url: String): VideoInfo?
     suspend fun put(url: String, videoInfo: VideoInfo)
+    suspend fun updateActualFormat(url: String, actualFormat: VideoInfo.Format)
 }